Refactor optimiser into modules
